+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - Header comment check to pr checklist
+- Explicit typing ESLint rule for useState
 - ESLint rule for TODO and FIXME comments
 
 ### Changed